HW #06: Second review fix
